Add status command

+++ internal/domain/status/status.go
@@ -3,14 +3,14 @@ package status
 import "github.com/bruli/raspberryRainSensor/pkg/common/vo"
 
 type Status struct {
-	humidity        float64
+	humidity        int
 	raining         bool
 	systemStartedAt vo.Time
-	temperature     float64
+	temperature     int
 	updatedAt       *vo.Time
 }
 
-func (s Status) Humidity() float64 {
+func (s Status) Humidity() int {
 	return s.humidity
 }
 
@@ -22,7 +22,7 @@ func (s Status) SystemStartedAt() vo.Time {
 	return s.systemStartedAt
 }
 
-func (s Status) Temperature() float64 {
+func (s Status) Temperature() int {
 	return s.temperature
 }
 
@@ -30,7 +30,7 @@ func (s Status) UpdatedAt() *vo.Time {
 	return s.updatedAt
 }
 
-func (s *Status) Hydrate(humidity float64, raining bool, startedAt vo.Time, temperature float64, updatedAt *vo.Time) {
+func (s *Status) Hydrate(humidity int, raining bool, startedAt vo.Time, temperature int, updatedAt *vo.Time) {
 	s.humidity = humidity
 	s.raining = raining
 	s.systemStartedAt = startedAt
